Extracted to Tasklist and Ui Class. Created Storage Class
