+++ src/main/java/Duke.java
@@ -1,192 +1,35 @@
 import java.util.Scanner;
-import java.util.ArrayList;
-import java.util.List;
+
+import exception.DukeException;
+import tasklist.*;
+import ui.Ui;
+import storage.Storage;
 
 public class Duke {
+    
+    private static taskList tasks = new taskList();
+    private static Ui userInterface;
+    private static Storage storage;
 
-    //private static Task [] taskList = new Task [100];
-    private static int taskCount = 0;
-
-    private static ArrayList<Task> taskList = new ArrayList();
-
-
-    public static void dukeGreet(){
-        System.out.println("    ________________________________________");
-        System.out.println ("    Hello! I'm Duke");
-        System.out.println ("    What can I do for you?");
-        System.out.println("    ________________________________________");
-    }
-
-    private static void dukeBye (){
-        System.out.println("    ________________________________________");
-        System.out.println("     Bye. Hope to see you again soon!");
-        System.out.println("    ________________________________________");
-    }
-
-/*
-      public static void dukeReply(Task[] taskList, int taskCount){
-        System.out.println("    ________________________________________");
-        System.out.println("     Got it. I've added this task: ");
-        System.out.println("     "+ taskList[taskCount-1].getDescription());
-        System.out.println("     Now you have " + taskCount + " tasks in the list.");
-        System.out.println("    ________________________________________");
-    }*/
-    public static void dukeReply(ArrayList<Task> taskList){
-        System.out.println("    ________________________________________");
-        System.out.println("     Got it. I've added this task: ");
-        System.out.println("     " + taskList.get(taskList.size()-1).getDescription());
-        System.out.println("     Now you have " + taskList.size() + " tasks in the list.");
-        System.out.println("    ________________________________________");
-}
-
-
-    public static void displayList (ArrayList<Task> tasks){
-        System.out.println("Here are the tasks in your list:");
-        for (int index = 0; index < tasks.size(); index++){
-            System.out.println (index+1 + "." + tasks.get(index).getDescription());
-        }
-    }
-
-    private static void markInList(String textInput) {
-        try {
-            System.out.println("    ________________________________________");
-            taskList.get(Integer.parseInt(textInput) - 1).markAsDone(true);
-            System.out.println("    Nice! I've marked this task as done: ");
-            System.out.println ("     " + taskList.get(Integer.parseInt(textInput) - 1).getDescription());
-            System.out.println("    ________________________________________");
-
-        } catch (NumberFormatException e){
-            //System.out.println ("☹ OOPS!!! You must indicate which task is done");
-        }
-
-    }
-
-    private static void deleteFromList(String textInput){
-        try {
-            System.out.println("    ________________________________________");
-            System.out.println("    Noted. I've removed this task: ");
-            System.out.println("     " + taskList.get(Integer.parseInt(textInput) - 1).getDescription());
-            taskList.remove(Integer.parseInt(textInput) - 1);
-            System.out.println("    Now you have " + taskList.size() + " tasks in the list.");
-            System.out.println("    ________________________________________");
-        } catch (NumberFormatException e){
-            //System.out.println ("☹ OOPS!!! You must indicate which task to delete");
-        }
-    }
-
-    public static void addTodo (Todo td){
-        taskList.add(td);
-        //taskCount++;
-        dukeReply(taskList);
-    }
-
-    public static void addDeadlines (Deadlines dl){
-        taskList.add(dl);
-        //taskCount++;
-        dukeReply(taskList);
-    }
-
-    public static void addEvent (Event e){
-        taskList.add(e);
-        //taskCount++;
-        dukeReply(taskList);
-    }
-
-    /*public static void displayList(Task[] tasks){
-        System.out.println("Here are the tasks in your list:");
-        for (int index = 0; index<taskCount; index++){
-            System.out.println (index+1 + "." + tasks[index].getDescription());
-        }
-    }
-
-     public static void addTodo (Todo td){
-        taskList[taskCount] = td;
-        taskCount++;
-        dukeReply(taskList, taskCount);
-    }
-
-    public static void addDeadlines (Deadlines dl){
-        taskList[taskCount] = dl;
-        taskCount++;
-        dukeReply(taskList, taskCount);
-    }
-
-    public static void addEvent (Event e){
-        taskList[taskCount] = e;
-        taskCount++;
-        dukeReply(taskList, taskCount);
-    }
-
-    private static void markInList(String textInput) {
-        try {
-            taskList[Integer.parseInt(textInput) - 1].markAsDone(true);
-            System.out.println("Nice! I've marked this task as done: ");
-            System.out.println (taskList[Integer.parseInt(textInput)-1].getDescription());
-
-        } catch (NumberFormatException e){
-            System.out.println ("☹ OOPS!!! You must indicate which task is done");
-        }
-
-    }*/
-
-
-    private static void dukeInput (String textInput) throws DukeException {
-        String textInputArr[] = textInput.split(" ",2);
-        try {
-            switch (textInputArr[0]) {
-                case "list":
-                    displayList(taskList);
-                    break;
-                case "done":
-                    markInList(textInputArr[1]);
-                    break;
-                case "delete":
-                    deleteFromList(textInputArr[1]);
-                    break;
-                case "todo":
-                    addTodo(new Todo(textInputArr[1]));
-                    break;
-                case "deadline":
-                    String textDeadline[] = textInputArr[1].split("/by", 2);
-                    if (textDeadline.length < 2) throw new DukeException("/by");
-                    addDeadlines(new Deadlines(textDeadline[0], textDeadline[1]));
-                    break;
-                case "event":
-                    String textEvent[] = textInputArr[1].split("/at", 2);
-                    if (textEvent.length < 2) throw new DukeException("/at");
-                    addEvent(new Event(textEvent[0], textEvent[1]));
-                    break;
-                case "bye":
-                    dukeBye();
-                    break;
-                default:
-                    throw new DukeException(textInputArr[1]);
-            }
-        } catch (IndexOutOfBoundsException e){
-            throw new DukeException(textInputArr[0]);
-        }
-        catch (DukeException e){
-            //do we need anything here
-        }
-
+    public Duke (String filePath){
+        userInterface = new Ui();
+        userInterface.showWelcome();
+        storage = new Storage(filePath);
     }
 
     public static void main(String[] args) {
-        String logo = " ____        _        \n"
-                + "|  _ \\ _   _| | _____ \n"
-                + "| | | | | | | |/ / _ \\\n"
-                + "| |_| | |_| |   <  __/\n"
-                + "|____/ \\__,_|_|\\_\\___|\n";
-        System.out.println("Hello from\n" + logo);
 
-        dukeGreet();
+        userInterface = new Ui();
+        userInterface.showWelcome();
+
+
         Scanner in = new Scanner(System.in);
         String textInput;
 
         do {
          textInput = in.nextLine();
           try {
-              dukeInput(textInput);
+              userInterface.dukeInput(tasks, textInput);
           } catch (DukeException e){
               //do we need anything here?
           }
