+++ src/main/java/tasklist/Deadlines.java
@@ -1,4 +1,8 @@
-public class Deadlines extends Task{
+package tasklist;
+
+import tasklist.Task;
+
+public class Deadlines extends Task {
     protected String by;
 
     public Deadlines(String description, String by){
