+++ src/main/java/tasklist/Task.java
@@ -1,3 +1,5 @@
+package tasklist;
+
 public class Task {
     protected String description;
     protected boolean isDone;
